arangement dossier pour fichier d'installation

+++ setup.sh
@@ -1,7 +1,17 @@
-# FIchier d'installation de la librairie
+#!/bin/bash
 
-sudo cp TR_Acquisition.py /usr/lib/python2.7 
-sudo cp TR_Acquisition.pyc /usr/lib/python2.7
-sudo cp Acquisition.py  /usr/lib/python2.7
-sudo cp Acquisition.pyc /usr/lib/python2.7
+if [ -e /usr/lib/python2.7 ]
+then
+
+  if [ $(uname --machine) == "x86_64" ]
+  then
+  # mettre installation de la bitlib
+  echo "Archi 64 bit détecté"
+  fi
+  sudo cp TR_Acquisition.py /usr/lib/python2.7 
+  sudo cp Acquisition.py  /usr/lib/python2.7
+else
+  echo "La version 2.7 de python est requise, elle ne semble pas être installé sur votre distribution linux"
+fi
+
